backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0
reportlab>=4.1,<5.0
gunicorn>=22.0,<23.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
whitenoise>=6.6.0
# MongoDB Support
django-mongodb-backend>=5.0.0
pymongo[srv]>=4.6.0
